extract model entity mapper object
