+++ apps/server/src/repositories/learnroom/dashboard.model.mapper.spec.ts
@@ -1,14 +1,15 @@
 import { ObjectId } from '@mikro-orm/mongodb';
 import { DashboardEntity } from '../../entities/learnroom/dashboard.entity';
-import { DashboardModelEntity, DashboardGridElementModel, mapToEntity, mapToModel } from './dashboard.model.entity';
+import { DashboardModelEntity } from './dashboard.model.entity';
+import { DashboardModelMapper } from './dashboard.model.mapper';
 
-describe('Dashboard Model Entity', () => {
+describe('Dashboard Model Entity Mapper', () => {
 	describe('mapToModel', () => {
 		it('transforms plain entity into modelEntity', () => {
 			const id = new ObjectId();
 			const entity = new DashboardEntity(id.toString(), { grid: [] });
 
-			const result = mapToModel(entity);
+			const result = DashboardModelMapper.mapToModel(entity);
 
 			expect(result instanceof DashboardModelEntity).toEqual(true);
 			expect(result._id).toEqual(id);
